consertando escala

+++ projetil.h
@@ -20,7 +20,7 @@ struct Projetil {
 typedef struct Projetil * projetil;
 
 /* Constructor para um projetil */
-projetil novo_projetil (double massa, double x, double y, double velx, double vely, double tempo, double inclinacao, WINDOW* w1);
+projetil novo_projetil (double massa, double x, double y, double velx, double vely, double tempo, int inclinacao, WINDOW* w1);
 
 /* Destructor para um projetil */
 void destroi_projetil (projetil p);
